treewide: Extract `hl` to `std` type converter to `conv::tc`.
